Move switchGamemode method into GamemodeSwitchHandler.kt
